creates a new container/ file and amends the text for phone vetting
