+++ i18n/src/i18n-messages.json
@@ -557,7 +557,7 @@
   },
   {
     "id": "letter.modal_confirm_info",
-    "defaultMessage": "The letter will contain a code that you ender here to verify your identity. The code sent to you will expire in 2 weeks starting from now"
+    "defaultMessage": "The letter will contain a code that you enter here to verify your identity. The code sent to you will expire in 2 weeks starting from now"
   },
   {
     "id": "letter.verify_title",
@@ -596,12 +596,12 @@
     "defaultMessage": "Go to Settings to confirm your number. You can only use this option with a confirmed number."
   },
   {
-    "id": "lmp.confirm_info",
-    "defaultMessage": "The phone operators are responsible for keeping this registry up to date. If you get an error despite knowing your number is valid, you might be missing and you need to choose another way to verify your identity."
-  },
-  {
     "id": "lmp.modal_confirm_title",
     "defaultMessage": "Check if your phone number is connected to your id number."
+  },
+  {
+    "id": "lmp.confirm_info",
+    "defaultMessage": "This check will be done in a registry updated by the phone operators. If they have not added your details, we won't be able to find you and you need to choose another way to verify your identity."
   },
   {
     "id": "lmp.verification_success",
