+++ src/containers/VerifyIdentity.js
@@ -1,0 +1,29 @@
+import { connect } from "react-redux";
+import VerifyIdentity from "components/VerifyIdentity";
+import i18n from "i18n-messages";
+
+const mapStateToProps = (state, props) => {
+  let verifiedNinStatus = "";
+  const nins = state.nins.nins.filter(nin => nin.verified);
+  nins.length >= 1 ? (verifiedNinStatus = true) : (verifiedNinStatus = false);
+  return {
+    nins: state.nins.nins,
+    verifiedNin: nins,
+    verifiedNinStatus: verifiedNinStatus,
+    is_configured: state.config.is_configured,
+    letter_verification: state.letter_proofing.confirmingLetter,
+    proofing_methods: state.config.PROOFING_METHODS,
+    message: state.nins.message
+  };
+};
+
+const mapDispatchToProps = (dispatch, props) => {
+  return {};
+};
+
+const VerifyIdentityContainer = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(VerifyIdentity);
+
+export default i18n(VerifyIdentityContainer);
